IAtom.hydrogenCount: int -> Integer
git-svn-id: https://cdk.svn.sourceforge.net/svnroot/cdk/trunk/cdk@8377 eb4e18e3-b210-0410-a6ab-dec725e4b171

+++ src/org/openscience/cdk/test/PseudoAtomTest.java
@@ -134,7 +134,7 @@ public class PseudoAtomTest extends CDKTestCase {
     public void testSetHydrogenCount_int() {
         IPseudoAtom atom = builder.newPseudoAtom("Whatever");
         atom.setHydrogenCount(+5);
-        assertEquals(0, atom.getHydrogenCount());
+        assertEquals(0, atom.getHydrogenCount().intValue());
     }
 
     public void testSetCharge_double() {
